refactor(data-model): remove enum re-exports and switch callers to direct enum modules
