+++ src/actors/rqg-actor-sheet-data-prep.async.test.ts
@@ -6,8 +6,8 @@ import {
   organizeEmbeddedItems,
 } from "./rqg-actor-sheet-data-prep";
 import { ItemTypeEnum } from "@item-model/item-types.ts";
-import { RuneTypeEnum } from "@item-model/rune-data-model.ts";
-import { SkillCategoryEnum } from "@item-model/skill-data-model.ts";
+import { RuneTypeEnum } from "@item-model/rune-enums.ts";
+import { SkillCategoryEnum } from "@item-model/skill-enums.ts";
 
 type ItemCollection = any[] & { get: (id: string) => any };
