+++ src/actors/rqg-actor-sheet-data-prep.ts
@@ -15,9 +15,9 @@ import type { GearItem } from "@item-model/gear-data-model.ts";
 import type { CultItem } from "@item-model/cult-data-model.ts";
 import type { SpiritMagicItem } from "@item-model/spirit-magic-data-model.ts";
 import type { SkillItem } from "@item-model/skill-data-model.ts";
-import { SkillCategoryEnum } from "@item-model/skill-data-model.ts";
+import { SkillCategoryEnum } from "@item-model/skill-enums.ts";
 import type { RuneItem } from "@item-model/rune-data-model.ts";
-import { RuneTypeEnum } from "@item-model/rune-data-model.ts";
+import { RuneTypeEnum } from "@item-model/rune-enums.ts";
 import type { HitLocationItem } from "@item-model/hit-location-data-model.ts";
 import type { UsageType, WeaponItem } from "@item-model/weapon-data-model.ts";
 import type { PassionItem } from "@item-model/passion-data-model.ts";
