+++ src/applications/improve-dialogs/improve-characteristic-dialog.ts
@@ -2,7 +2,8 @@ import { RqgActor } from "../../actors/rqg-actor";
 import { convertFormValueToString, isDocumentSubType, localize, RqgError } from "../../system/util";
 import { systemId } from "../../system/config";
 import { ItemTypeEnum } from "@item-model/item-types.ts";
-import { type CultItem, CultRankEnum } from "@item-model/cult-data-model.ts";
+import type { CultItem } from "@item-model/cult-data-model.ts";
+import { CultRankEnum } from "@item-model/cult-enums.ts";
 import { templatePaths } from "../../system/load-handlebars-templates";
 import type { Characteristics } from "../../data-model/actor-data/characteristics";
 import { RqgLogger } from "../../system/logging/rqg-logger";
