+++ src/items/cult-item/cult-sheet-v2.ts
@@ -1,4 +1,5 @@
-import { CultRankEnum, type CultItem } from "@item-model/cult-data-model.ts";
+import type { CultItem } from "@item-model/cult-data-model.ts";
+import { CultRankEnum } from "@item-model/cult-enums.ts";
 import {
   isTruthy,
   getRequiredDomDataset,
