+++ src/items/cult-item/cult-sheet.ts
@@ -1,5 +1,6 @@
 import { ItemTypeEnum } from "@item-model/item-types.ts";
-import { CultRankEnum, type CultItem } from "@item-model/cult-data-model.ts";
+import type { CultItem } from "@item-model/cult-data-model.ts";
+import { CultRankEnum } from "@item-model/cult-enums.ts";
 import {
   isTruthy,
   getRequiredDomDataset,
