+++ src/items/hit-location-item/hit-location-sheet.ts
@@ -1,10 +1,10 @@
 import { ItemTypeEnum } from "@item-model/item-types.ts";
 import {
-  type HitLocationHealthState,
   hitLocationHealthStatusOptions,
   type HitLocationItem,
-  HitLocationTypesEnum,
 } from "@item-model/hit-location-data-model.ts";
+import type { HitLocationHealthState } from "@item-model/hit-location-enums.ts";
+import { HitLocationTypesEnum } from "@item-model/hit-location-enums.ts";
 import { getSelectHitLocationOptions } from "../../system/util";
 import { RqgItemSheet } from "../rqg-item-sheet";
 import { systemId } from "../../system/config";
