+++ src/items/rqg-item.ts
@@ -42,7 +42,8 @@ import { HitLocationDataModel } from "@item-model/hit-location-data-model";
 import { HomelandDataModel } from "@item-model/homeland-data-model";
 import { OccupationDataModel } from "@item-model/occupation-data-model";
 import { type SpellItem, spellItemTypes } from "@item-model/spell.ts";
-import type { DamageType, UsageType, WeaponItem } from "@item-model/weapon-data-model.ts";
+import type { UsageType, WeaponItem } from "@item-model/weapon-data-model.ts";
+import type { DamageType } from "@item-model/weapon-enums.ts";
 import { DamageDegree } from "../system/combat-calculations.defs";
 import { dodgeBaseChance, jumpBaseChance } from "./skill-item/skill-formulas";
 import type { RuneMagicItem } from "@item-model/rune-magic-data-model.ts";
