+++ src/items/rune-item/rune-sheet.ts
@@ -1,5 +1,6 @@
 import { ItemTypeEnum } from "@item-model/item-types.ts";
-import { RuneTypeEnum, type RuneItem } from "@item-model/rune-data-model.ts";
+import type { RuneItem } from "@item-model/rune-data-model.ts";
+import { RuneTypeEnum } from "@item-model/rune-enums.ts";
 import { localize, getSelectRuneOptions } from "../../system/util";
 import { RqgItemSheet } from "../rqg-item-sheet";
 import { systemId } from "../../system/config";
